Separate lexer and parser modules
