+++ src-tauri/src/dqmj1_rom/events/assembly/mod.rs
@@ -1,0 +1,2 @@
+pub mod lexer;
+pub mod parser;
